fix(LibCarla): eliminate all 282 compiler warnings. (#9574)
* fix(LibCarla): resolve all 282 compiler warnings across 27 files

Mechanical, zero-behavior-change refactoring that eliminates every
compiler warning emitted during LibCarla server and client builds.
Each warning category corresponds to a concrete defect class: silent
integer truncation on large maps, NRVO-preventing moves, deep copies in
parser loops, discarded ROS2 command responses, and C-style casts that
bypass type checking. Fixing them removes the noise that masks real issues
and establishes the baseline needed to enforce -Werror in CI.

Changes by category:
- Fix misleading indentation in FIleTransfer.cpp (brace correction)
- Eliminate unnecessary deep-copies in range-for loops (parsers)
- Remove pessimizing std::move() on return statements (NRVO)
- Reorder constructor initializer lists to match declaration order
- Remove unused variables, lambda captures, and dead code
- Annotate unused parameters with /*name*/ comment syntax (C++14)
- Replace implicit signed/unsigned conversions with size_t or static_cast
- Replace implicit double-to-float with explicit static_cast<float> or float literals
- Replace C-style casts with static_cast / reinterpret_cast
- Fix third-party headers: namespace pollution, bitfield sign, duplicate macros

Files changed (27):
  carla/Exception.cpp
  carla/client/FIleTransfer.cpp
  carla/client/detail/Simulator.cpp
  carla/geom/Mesh.cpp
  carla/geom/Simplification.cpp
  carla/multigpu/primary.cpp
  carla/multigpu/primaryCommands.cpp
  carla/multigpu/router.cpp
  carla/multigpu/secondary.cpp
  carla/opendrive/parser/GeoReferenceParser.cpp
  carla/opendrive/parser/GeometryParser.cpp
  carla/opendrive/parser/RoadParser.cpp
  carla/road/Deformation.h
  carla/road/Map.cpp
  carla/road/MapBuilder.cpp
  carla/road/MeshFactory.cpp
  carla/road/element/Geometry.cpp
  carla/trafficmanager/ALSM.cpp
  carla/trafficmanager/LocalizationStage.cpp
  test/client/test_image.cpp
  test/client/test_opendrive.cpp
  test/common/test_geom.cpp
  test/server/test_benchmark_streaming.cpp
  third-party/marchingcube/Cube.h
  third-party/marchingcube/MeshReconstruction.h
  third-party/marchingcube/Triangulation.h
  third-party/simplify/Simplify.h

Result: make LibCarla.server.release and make LibCarla.client.release
produce 0 warnings from LibCarla code. 44/45 unit tests pass;
the single failure (vector3D.make_unit_vec) is pre-existing.

* Added log record of the PR changes to CHANGELOG.md

+++ LibCarla/source/test/client/test_image.cpp
@@ -80,7 +80,7 @@ TEST(image, depth) {
           get_color(p_bgra8, red_t()) = static_cast<uint8_t>(r);
           get_color(p_bgra8, green_t()) = static_cast<uint8_t>(g);
           get_color(p_bgra8, blue_t()) = static_cast<uint8_t>(b);
-          const float depth = r + (g * 256) + (b  * 256 * 256);
+          const float depth = static_cast<float>(r + (g * 256) + (b  * 256 * 256));
           const float normalized = depth / static_cast<float>(256 * 256 * 256 - 1);
           p_gray8[0] = static_cast<uint8_t>(255.0 * normalized);
